EXTEND-8 : Add a listener service to manage extendable resource removal. Add an implementation of listener to remove extenders of removed resources.
git-svn-id: http://dev.lutece.paris.fr/svn/lutece/portal/trunk/plugins/extends/module-extend-comment@48511 bab10101-e421-0410-a517-8ce0973de3ef

+++ src/java/fr/paris/lutece/plugins/extend/modules/comment/service/extender/CommentResourceExtender.java
@@ -123,7 +123,10 @@ public class CommentResourceExtender extends AbstractResourceExtender
     @Override
     public void doDeleteResourceAddOn( ResourceExtenderDTO extender )
     {
-        _configService.remove( extender.getIdExtender(  ) );
+        if ( extender.getIdExtender( ) > 0 )
+        {
+            _configService.remove( extender.getIdExtender( ) );
+        }
         _commentService.removeByResource( extender.getIdExtendableResource( ), extender.getExtendableResourceType( ) );
     }
 }
